+++ Aprendizagem Industrial Programador de Sistemas da Informação/DESENVOLVIMENTO DE SISTEMAS/Spring/escola/src/main/java/com/psii/escola/service/EstudanteService.java
@@ -2,6 +2,8 @@ package com.psii.escola.service;
 
 import com.psii.escola.model.Estudante;
 import com.psii.escola.repository.EstudanteRepository;
+import com.psii.escola.repository.InscricaoRepository; // Importando o repositório de Inscrição
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,19 +15,27 @@ public class EstudanteService {
     @Autowired
     private EstudanteRepository estudanteRepository;
 
+    @Autowired
+    private InscricaoRepository inscricaoRepository; // Injeção do repositório de Inscrição
+
     public List<Estudante> getAllEstudantes() {
         return estudanteRepository.findAll();
     }
 
     public Estudante getEstudanteById(Long id) {
-        return estudanteRepository.findById(id).orElse(null);
+        return estudanteRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Estudante não encontrado com id: " + id));
     }
 
     public void saveEstudante(Estudante estudante) {
         estudanteRepository.save(estudante);
     }
 
-    public void deleteEstudante(Long id) {
-        estudanteRepository.deleteById(id);
+    public void excluirEstudante(Long id) {
+        // Verifique se o estudante existe antes de tentar excluir
+        if (!estudanteRepository.existsById(id)) {
+            throw new IllegalStateException("Estudante com ID " + id + " não encontrado.");
+        }
+        estudanteRepository.deleteById(id); // Exclui o estudante do banco de dados
     }
 }
